backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
motor==3.7.1
docker==7.1.0
pydantic==2.12.5
python-dotenv==1.2.2
websockets==16.0
httpx==0.28.1
